fmp4mux: Language and orientation are stream tags
The language tag is purely a stream specific tag and the orientation
tag can be both, stream or global scope at the same time.

Language tags in global scope are considered to be stream scope.

Orientation tags are also set in TKHD but combined with the global
orientation tag if available.

And switched to GStreamer API for image orientation tag storage and
handling and added the flip transformations.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-rs/-/merge_requests/2033>