net/aws: consolidate apostrophes with next word in span_tokenize_items
As the text to translate in span-based mode looks like:

<span>the</span> <span>year</span>

We may receive back from the service when translating to French:

<span>l'</span> <span>année</span>

Which in turn means we'll push out two items. It makes more sense to
push those as a single item, as this will prevent downstream elements
from inserting a space.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-rs/-/merge_requests/2126>