mux: use smaller number of samples for testing
This improves the test time from ~30 seconds to ~5 seconds, mainly
because of the AV1 encoding. However reduction in the uncompressed
tests is also worth while because there are a lot of them, and
will likely be more as the number of formats supported on both
encode side and in qtdemux for round-trip.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-rs/-/merge_requests/2130>