cea608tott: make test less shocking by having more cues as context
The contents of the current test for cea608tott can be a bit shocking
without context about the source.

This patch adds more cues to the test so that a reader just trying to
understand the elements has more context that this is a news program.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-rs/-/merge_requests/2249>