meson, cargo: don't build skia plugin by default for now
Fails to build in some/many cases:

https://gitlab.freedesktop.org/gstreamer/gst-plugins-rs/-/issues/692
https://gitlab.freedesktop.org/gstreamer/cerbero/-/merge_requests/1804

In addition to the git-sync-deps issue, the skia-safe project uses
cargo to invoke gn, and in the process it loses the ability to pick up
dependencies from the system. This means it cannot integrate as
a meson subproject.

Co-authored-by: L. E. Segovia <amy@amyspark.me>
Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-rs/-/merge_requests/2251>