webrtcsink: Don't require encoder element for pre-encoded streams
When webrtcsink takes as an input a stream that is already encoded, it
errors out when there is no encoder element available for the codec in
question.

This change makes it possible to stream an output from a camera that
already produces e.g. video/x-h264 without bloating the GStreamer
installation with a redundant H.264 encoder element.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-rs/-/merge_requests/2273>