net: hlssink3: Use closed fragment location in playlist generation
Currently, we incorrectly use the last location we generated, which is
the current in-flight fragment while generating a new playlist. Instead,
the playlist should use the fragment that was just close (i.e. the
previous fragment), so it does not refer to a fragment we have not yet
uploaded.

This is a bug that was copied over from how hlssink2 does things.

Fixes: https://gitlab.freedesktop.org/gstreamer/gst-plugins-rs/-/issues/701
Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-rs/-/merge_requests/2324>