hlscmafsink: Add support for I-frame only playlist
There are two ways of doing I-frame only playlists.
- Generate a playlist with fragments that only contain I-frames,
  with the stream possibly being encoded I-frame at a different
  frame rate.
- Write out a manifest/playlist that just has the key frames with
  byte ranges into the original fragments.

We opt for the second approach here. This I-frame playlist will
only be generated along side the media playlist if the relevant
property is set. Note that this is different to i-frames-only in
hlssink3 which generates an I-frame playlist using the first
approach.

The Cargo.toml changes are required for accessing the meta.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-rs/-/merge_requests/2366>