textaccumulate: expose extend-duration property
When set to true, the element will extend the duration of the last
currently-draining item to the start time of the next not-yet-drained
item, if any.

A second property is exposed to allow controlling the duration of the
gap to preserve with the next item.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-rs/-/merge_requests/2564>