gtk4: Implement cropped imports without viewport
This fixes glitches with padded buffers, notably from V4L2 stateless
decoders when used with codecs like VP9 and HEVC, as well as import
errors for subsampled buffers with odd sizes.

The later can be tested with:
gst-launch-1.0 videotestsrc ! video/x-raw,format=I420,width=639,height=479 ! gtk4paintablesink

Note that the approach taken here does *not* use viewports with the full
vmeta size. That is used in waylandsink and gtkgstsink and apparently
sometimes is needed with certain formats/modifiers. This commit is
simpler, but solves the mentioned issues on all devices that I could
test so far and thus a good intermediate step.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-rs/-/merge_requests/2648>