threadshare: runtime task: execute action in downward transition
... even if the Task is in the Error state.

This is needed to clean things up. Upward transitions from the error state must
fail, however a `TaskImpl` must be able to execute the downward state transition
action and the resulting transition should not bubble up as an error to the
element.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-rs/-/merge_requests/2602>