fallbacksrc: Add support for encoded outputs
fallbacksrc uses uridecodebin3 for primary and fallback uri unless a
custom source is used. Support encoded outputs by allowing to set the
caps property on uridecodebin3.

We introduce a video-caps & audio-caps property which are then merged
and set on the `caps` property for uridecodebin3 for both the primary
and fallback uri. These are set to <video/audio>/x-raw by default and
do not affect the existing functionality unless set explicitly.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-rs/-/merge_requests/2591>