threadshare: udpsink: fix panic recalculating latency from certain executors
Depending on the executor, recalculating latency could lead to a panic:

> cannot execute `LocalPool` executor from within another executor

This is solved by using `block_on_or_add_subtask()` for the sync wrappers using
the async Mutex which are mostly setters or modifiers.

Special case for `clients()`, which is a getter used to read the "clients"
property. The getter is removed so as to avoid using it in other places where we
could use a non problematic solution instead. We keep using `block_on` at the
call site for now. If it proves limiting, we could synchronize a copy of
`clients` in the `settings` upon modifications, so we don't have to go through
`sink_pad_handler` to read its value.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-rs/-/merge_requests/2609>