speechmaticstranscriber: in-depth refactoring
This transcriber element had several design issues that needed
refactoring:

Segment handling was deficient: the element attempted to track a DISCONT
offset in order to keep a single connection and zero-based output segment.

This didn't work very well and caused some complex and unorthodox code.

Instead, drain on new segments (and DISCONT buffers), close the
connection and re-open it later if necessary. Input segments are
forwarded, and output timestamps now match the input timestamps.

Gap generation was too complex: the element implemented a timeout logic
on all of its source pads to guarantee a steady stream of gap events
output on time.

I have decided that it is sufficient to simply push gap events as we
receive (potentially empty) transcripts, this is sufficient as long as
the delay from the speechmatics API remains under the configured
latency.

In order to help the application ensure that, a new `max-observed-delay`
readonly property is exposed, and a warning message is posted whenever
the observed delay exceeds the configured latency (+ lateness).

A related change is that the element also no longer shifts forward the
timestamps of late items. It was a good-intentioned solution that ended
up causing more pernicious issues to debug further downstream.

The element no longer syncs its input to the clock.

The `enable-late-punctuation-hack` property is now deprecated and
inoperant, as the hack no longer seems useful in my testing.

A SECTION level documentation for the element has been added to provide
some recommendations for using it.

Unsynced pads, which had been deprecated earlier by the posting of
messages, are now removed.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-rs/-/merge_requests/2617>