net/quinn: Support sharing of QUIC/WebTransport connection/session
This adds support for sharing of a QUIC/WebTransport connection/session
with an element upstream or downstream. This is required for supporting
Media over QUIC (MoQ).

MoQ has the notion of control channel which is a bi-directional stream
and data flows over uni-directional streams. We want to reuse QUIC and
WebTransport elements for the same. Bi-directional streams however do
not make sense in the context of a sink or source element.

We use Context to share the QUIC connection or WebTransport session,
with which the element needing the bi-directional stream can choose
to open the required control channel. Rest of the functionality in
either of the source or sink element is not changed.

Below is an example of how this would work for QUIC/WT sink with MoQ.

     Bi-directional stream (opened using shared Context)
   --------------------------------------------------------|
   |                                                       |
   |                                                       |
 -------        Context       -------   QUIC/WT       -----------
  moqmux --------------------  sink   ------------->   MoQ Relay
 ------- (Connection/Session) ------- Unidirectional  -----------

Below is an example of how this would work for QUIC/WT source with MoQ.

     Bi-directional stream (opened using shared Context)
   ----------------------------------------------------------------|
   |                                                               |
   |                                                               |
 -----------       QUIC/WT       -------        Context        -----------
  MoQ Relay  <------------------ source  ---------------------  moqdemux
 -----------    Unidirectional   -------  (Connection/Session) -----------

Data only ever flows on the unidirectional streams above and these are
solely under the purview of sink and source. Connection/Session itself
is also only ever initiated by the sink or source.

We additionally introduce role for quinnwtserversink as when used for
MoQ publishing, it will have to initiate the connection with relay.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-rs/-/merge_requests/2643>