net/quinn: Rename quinnwtserversink to quinnwtsink
For Media Over QUIC, we need the `role` property so that the sink can
initiate the session establishment. As this can get confusing, rename
the element to `quinnwtsink` as `quinnwtserversink` implies that the
sink would only listens for incoming session establishment request.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-rs/-/merge_requests/2643>