whipserversrc: implement example for external HTTP server
The current implementation of the signaller will always spawn a
web server and expose multiple sessions from the same webrtcsrc
element.

This is not very ergonomic, as it forces the application that
wants to spawn a whip server to ingest multiple sessions in the
same pipeline.

Instead it is preferable to have the application spawn the web
server itself and create one webrtcsrc per session, then relay
requests and responses to the signaller.

For this purpose, new signals are exposed on the WhipServerSignaller
object (post, delete).

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-rs/-/merge_requests/2709>