whepserversink: implement example for external HTTP server
The current implementation of the signaller will always spawn a
web server.

An application may wish instead to spawn its own http server.

Instead it is desirable to allow the application to spawn the web
server itself then relay requests and responses to the signaller.

For this purpose, new signals are exposed on the WhepServerSignaller
object (post, delete).

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-rs/-/merge_requests/2733>